+++ quic/s2n-quic-transport/src/space/application.rs
@@ -166,7 +166,7 @@ impl<Config: connection::Config> ApplicationSpace<Config> {
         self.ack_manager.largest_received_packet_number_acked()
     }
 
-    pub fn bytes_in_flight(&self) -> u64 {
+    pub fn bytes_in_flight(&self) -> usize {
         self.recovery_manager.bytes_in_flight()
     }
